refactor: Rename swiper-bundle
