+++ css/style.css
@@ -11,7 +11,7 @@
 @import url(reset.css);
 @import url(contacts.css);
 @import url(footer.css);
-@import url(swiper-bundle.min.css);
+@import url(swiper-bundle.8.4.5.min.css);
 
 
 body{
