+++ js/modules/swiperCommentsFunctionality.js
@@ -1,5 +1,5 @@
 import { setSlidesState } from "./swiperHelpers.js";
-import Swiper from "./swiper-bundle.esm.browser.min.js";
+import Swiper from "./swiper-bundle.8.4.5.esm.browser.min.js";
 
 const BTN_APPEARANCE_DELAY = 1000;
 const RESIZE_DELAY = 250;
